+++ frontend/table.js
@@ -34,6 +34,7 @@ class Table {
     }
     this.columns = columns
     this.entries = Object.freeze({})
+    this.actorIds = Object.freeze({})
     Object.freeze(this)
   }
 
@@ -42,6 +43,13 @@ class Table {
    */
   byId(id) {
     return this.entries[id]
+  }
+
+  /**
+   * Returns the ID of the actor that added the row with object ID `id`.
+   */
+  getActorId(id) {
+    return this.actorIds[id]
   }
 
   /**
@@ -145,17 +153,19 @@ class Table {
     if (!this[OBJECT_ID]) {
       throw new RangeError('clone() requires the objectId to be set')
     }
-    return instantiateTable(this[OBJECT_ID], copyObject(this.entries))
+    return instantiateTable(this[OBJECT_ID], copyObject(this.entries), copyObject(this.actorIds))
   }
 
   /**
-   * Sets the entry with key `id` to `value`.
+   * Sets the entry with key `id` to `value`. `actorId` is the ID of the actor
+   * that performed this assignment.
    */
-  set(id, value) {
+  set(id, value, actorId) {
     if (Object.isFrozen(this.entries)) {
       throw new Error('A table can only be modified in a change function')
     }
     this.entries[id] = value
+    this.actorIds[id] = actorId
     if (id === 'columns') this.columns = value
   }
 
@@ -167,6 +177,7 @@ class Table {
       throw new Error('A table can only be modified in a change function')
     }
     delete this.entries[id]
+    delete this.actorIds[id]
   }
 
   /**
@@ -218,7 +229,8 @@ class WriteableTable extends Table {
    */
   get columns() {
     const columnsId = this.entries.columns[OBJECT_ID]
-    return this.context.instantiateObject(this.path, columnsId) // FIXME add 'columns' element to path
+    const path = this.path.concat([{key: 'columns', objectId: columnsId}])
+    return this.context.instantiateObject(path, columnsId)
   }
 
   /**
@@ -227,7 +239,8 @@ class WriteableTable extends Table {
    */
   byId(id) {
     if (isObject(this.entries[id]) && this.entries[id][OBJECT_ID] === id) {
-      return this.context.instantiateObject(this.path, id) // FIXME add id element to path
+      const path = this.path.concat([{key: id, objectId: id}])
+      return this.context.instantiateObject(path, id)
     }
   }
 
@@ -265,11 +278,15 @@ class WriteableTable extends Table {
  * This function is used to instantiate a Table object in the context of
  * applying a patch (see apply_patch.js).
  */
-function instantiateTable(objectId, entries) {
+function instantiateTable(objectId, entries, actorIds) {
   const instance = Object.create(Table.prototype)
+  if (!objectId) {
+    throw new RangeError('instantiateTable requires an objectId to be given')
+  }
   instance[OBJECT_ID] = objectId
   instance[CONFLICTS] = Object.freeze({})
   instance.entries = entries || {}
+  instance.actorIds = actorIds || {}
   return instance
 }